Move point in multiple windows
With this commit, we take into account if point is near the bottom of
multiple windows.  Logic for moving point after the action was also
simplified; if point is moved while mini-buffer is active, we no
longer take the new point position into account.  Maybe this will have
to be improved later on.